+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ### Removed
 
 ### Fixed
+- Compatibility issue with Gradle < 9.0 (#46)
 - Ignored dependencies switch not taken into account in Gradle plugin (#47)
 
 ### Security